añadido metodo registrar, nueva migracion, cambios auxiliares en validadores y repositorios

+++ API/Core/Entidades/Movimiento.cs
@@ -8,7 +8,7 @@
 		public DateTime Fecha { get; set; }
 		public int TipoMovimientoId { get; set; }
 		public virtual TipoMovimiento? TipoMovimiento { get; set; }
-		public int CuentaOrigenIdentificador { get; set; }
-		public int CuentaReceptoraIdentificador { get; set; }
+		public Int64 CuentaOrigenIdentificador { get; set; }
+		public Int64 CuentaReceptoraIdentificador { get; set; }
 	}
 }
